Share Agda adapter parsing for migration tooling
